refactor:PuppyCatAnimationのAnimateを分離
